requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.1
Flask-RESTful==0.3.9
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.2.2
google-auth==2.3.2
google-cloud-core==2.2.1
google-cloud-datastore==2.4.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
libcst==0.3.21
MarkupSafe==2.0.1
mypy-extensions==0.4.3
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
PyGithub==1.55
python-dateutil==2.8.2
pytest==6.2.5
google-cloud-logging==2.7.0
coverage==6.2
